XMPP# core remaster.
In order to incorporate client and component connections into the library, I had to make the decision to merge "three into one".

Basically, the library will now use the expat parser by default to read XMPP; it is practical, fast and optimized. It requires the native library installed on the system (since it depends on PInvoking).

In the meantime... I have readjusted some classes so that the management of elements, attributes and nodes is more simplified (no more complex hacks!).

The XMPP# Tokenizer and the previous version of XMPP# Expat package will still be available for use. After all, the tokenizer has a simplified version of the SAX-style parser.

In many internal tests, everything is working as expected, including the "Simple Server" project continues working fine.